feat(board): UI layout for board content (no API) part 2 (separate into components)
