+++ src/pages/Boards/BoardContent/ColumnList/Column/CardList/CardList.jsx
@@ -1,0 +1,40 @@
+import Box from '@mui/material/Box'
+import Card from './Card/Card'
+
+function CardList() {
+  return (
+    <Box
+      sx={{
+        p: '0 5px',
+        m: '0 5px',
+        display: 'flex',
+        flexDirection: 'column',
+        gap: 1,
+        overflowX: 'hidden',
+        overflowY: 'auto',
+        maxHeight: (theme) => `calc(
+             ${theme.custom.boardContentHeight} -
+             ${theme.custom.columnHeaderHeight} -
+             ${theme.custom.columnfooterHeight} -
+             ${theme.spacing(5)}
+            )`,
+        '&::-webkit-scrollbar-thumb': {
+          backgroundColor: '#ced0da'
+        },
+        '&::-webkit-scrollbar-thumb:hover': {
+          backgroundColor: '#bfc2cf'
+        }
+      }}
+    >
+      <Card />
+      <Card temporaryHideMedia />
+      <Card temporaryHideMedia />
+      <Card temporaryHideMedia />
+      <Card temporaryHideMedia />
+      <Card temporaryHideMedia />
+      <Card temporaryHideMedia />
+    </Box>
+  )
+}
+
+export default CardList
