+++ src/pages/Boards/BoardContent/ColumnList/Column/Column.jsx
@@ -1,0 +1,157 @@
+import Box from '@mui/material/Box'
+import Button from '@mui/material/Button'
+import Divider from '@mui/material/Divider'
+import ListItemIcon from '@mui/material/ListItemIcon'
+import ListItemText from '@mui/material/ListItemText'
+import Menu from '@mui/material/Menu'
+import MenuItem from '@mui/material/MenuItem'
+import Tooltip from '@mui/material/Tooltip'
+import Typography from '@mui/material/Typography'
+
+import AddCard from '@mui/icons-material/AddCard'
+import Cloud from '@mui/icons-material/Cloud'
+import ContentCopy from '@mui/icons-material/ContentCopy'
+import ContentCut from '@mui/icons-material/ContentCut'
+import ContentPaste from '@mui/icons-material/ContentPaste'
+import DeleteForeverIcon from '@mui/icons-material/DeleteForever'
+import DragHandle from '@mui/icons-material/DragHandle'
+import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
+
+import { useState } from 'react'
+import CardList from './CardList/CardList'
+
+function Column() {
+  const [anchorEl, setAnchorEl] = useState(null)
+  const open = Boolean(anchorEl)
+
+  const handleClick = (event) => {
+    setAnchorEl(event.currentTarget)
+  }
+  const handleClose = () => {
+    setAnchorEl(null)
+  }
+  return (
+    <Box
+      sx={{
+        minWidth: '300px',
+        maxWidth: '300px',
+        bgcolor: (theme) =>
+          theme.palette.mode === 'dark' ? '#333643' : '#ebecf0',
+        marginLeft: 2,
+        borderRadius: '6px',
+        height: 'fit-content',
+        maxHeight: (theme) =>
+          `calc(${theme.custom.boardContentHeight} - ${theme.spacing(5)})`
+      }}
+    >
+      {/*---- Header ----*/}
+      <Box
+        sx={{
+          height: (theme) => theme.custom.columnHeaderHeight,
+          display: 'flex',
+          justifyContent: 'space-between',
+          alignItems: 'center',
+          p: 2
+        }}
+      >
+        <Typography
+          variant="h6"
+          sx={{
+            fontSize: '1rem',
+            fontWeight: 'bold',
+            cursor: 'pointer',
+            color: (theme) =>
+              theme.palette.mode === 'dark'
+                ? 'white'
+                : theme.palette.primary.main
+          }}
+        >
+          Column Title
+        </Typography>
+        {/* options  */}
+        <Box>
+          <Tooltip title="More options">
+            <ExpandMoreIcon
+              sx={{ color: 'text.primary', cursor: 'pointer' }}
+              id="basic-column-dropdown"
+              aria-controls={open ? 'basic-menu-column-dropdown' : undefined}
+              aria-haspopup="true"
+              aria-expanded={open ? 'true' : undefined}
+              onClick={handleClick}
+            />
+          </Tooltip>
+
+          <Menu
+            id="basic-menu-column-dropdown"
+            anchorEl={anchorEl}
+            open={open}
+            onClose={handleClose}
+            slotProps={{
+              list: {
+                'aria-labelledby': 'basic-column-dropdown'
+              }
+            }}
+          >
+            <MenuItem>
+              <ListItemIcon>
+                <AddCard fontSize="small" />
+              </ListItemIcon>
+              <ListItemText>Add new card</ListItemText>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <ContentCut fontSize="small" />
+              </ListItemIcon>
+              <ListItemText>Cut</ListItemText>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <ContentCopy fontSize="small" />
+              </ListItemIcon>
+              <ListItemText>Copy</ListItemText>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <ContentPaste fontSize="small" />
+              </ListItemIcon>
+              <ListItemText>Paste</ListItemText>
+            </MenuItem>
+            <Divider />
+            <MenuItem>
+              <ListItemIcon>
+                <DeleteForeverIcon fontSize="small" />
+              </ListItemIcon>
+              <ListItemText>Remove this column</ListItemText>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <Cloud fontSize="small" />
+              </ListItemIcon>
+              <ListItemText>Archive this column</ListItemText>
+            </MenuItem>
+          </Menu>
+        </Box>
+      </Box>
+      {/* Card list */}
+      <CardList />
+
+      {/*---- Footer ----*/}
+      <Box
+        sx={{
+          height: (theme) => theme.custom.columnFooterHeight,
+          display: 'flex',
+          alignItems: 'center',
+          justifyContent: 'space-between',
+          p: 2
+        }}
+      >
+        <Button startIcon={<AddCard />}>Add new card</Button>
+        <Tooltip title="Drag to move">
+          <DragHandle sx={{ cursor: 'pointer' }} />
+        </Tooltip>
+      </Box>
+    </Box>
+  )
+}
+
+export default Column
